Add Docker files and scripts
By running "docker-compose -f docker/docker-compose.yml up" the frontend 
and a mock backend can be run. This enables unit and e2e testing.

There is also a Dockerfile for producing a production ready container 
using nginx as base.

The specs.yml file serves as an API definition for the backend mock 
container. This needs to be updated (i.e. copied over) manually whenever 
the backend API changes. This workflow may be improved in the future.