fixes a bug which made tests in modes without saveReponses to write back some data when test with saveReponses was opened before.
Explanation: The testmode comes later with getTestData and is set correctly in the RUNNING stage, while in LOADING stage it war still the testmode of the test before! So if we opened a test in TRIAL mode after running a test in HOT mode, the loading stage will be write back to backend. As a workaround, we reset to DEMO here, so while in loading stage `saveResponses` is always false.

As all tests of one login are always in the same mode, maybe it would be better to retrieve the testmode from the login instead by getTestData.