replaces the session if the data.testId
In the frist implementation of the group-monitor the test-sessions replaced a combination of person and test so they needed there own id. Now, they represent only tests and we don't show a person which is a logged in but has not started a test right now anymore, we don'T need the combined id anymore as well.