sends the `resume` command to every test-session, not only those, which are paused.
Why? Consider the following scenario before this change: we have 10 test-sessions prepared but 7 testees are present. The autoCheckAll-mode is on. We click pause and `pause` is sent to all ten test-sessions. The seven session which are online go in the pause mode and the controller sends back the state `{CONTROLLER: paused}`. Now we hit `resume` - and reach the only seven test-sessions which are in pause now. Now one of the the remaining three testees shows up and starts his session. Ge gets the pending pause from before command but no resume. Now he is paused, while the rest isn't.

To avoid this case of inconsistency we send the resume command to all test-sessions anyway.

Another approach would hve been to drop the command-"caching" for pending tests. But this would have other impacts in scenarios where testees who start later.