Add environment for e2e testing
The backend address in the environment file needs to either be localhost 
for using the app from the (host system's) browser of the internal 
docker name (backend-mock) so the frontend container can reach it. SO 
the testcenterUrl can either be reached from the host or fromt he 
frontend container but not both. This makes it so, that the e2e tests 
use a specific environment (defined in angular.json) which uses a 
different environment file for running e2e test within the frontend 
container.