Refactoring of the File-Upload-component
That was necessary because the dependency injection was changed in teh new angular version. The encapsulated upload-module got it's own instance of http client und thus the interceptor did not kick in (even when http was injected with @SkipSelf).

Tidy up the upload component, add types etc.

Does not work properly yet.