Prohibit infinite loops on error in testcontroller.
Circular subscriptions made the TestController raise infinite error messages on certain errors. To prohibit this, some refactoring was necessary. Now the TestStatus$.subscriptions does only logging, but has no other logic in it. Navigation to error (= status) page does not happen as  reaction to controllerState$.next(ERROR) anymore, but as reaction to mds.appError$.next.

The logging only takes place if the status was actually changed (in opposition to "the same status was pushed again"). Otherwise,  when the credentials would not work anymore for any reason, the subsequent error would be followed by another because the state change to ERROR could not be reported. This would lead to the next chnage of state and so on.

Also the state-page was refactored to be more readable.