Fix wicked bug: After an error, no new test was startable.
Reason: The ReplaySubject mds.appError$ replayed the last error(s) when it got subscribed on initialization of TestControllerComponent. Then it tried to navigate to the error-page (again), belonging to the last opened test (because mds.testId was not updated yet). So although it seemed to be a routing problem it turned out to be an error-handling problem.
Solution: After clicking on th error, it gets removed now by pushing <void> to mds.appError$. In addition, the replaySubject can only replay one error now.